+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changes
 
 - Add support for Python 3.7
 
+- Fix ``verifyObject`` for class objects with staticmethods on
+  Python 3. See `issue 126
+  <https://github.com/zopefoundation/zope.interface/issues/126>`_.
 
 4.5.0 (2018-04-19)
 ------------------